requirements.txt
Django>=4.2,<5.0
tensorflow>=2.10
python-chess>=1.0
djangorestframework>=3.14
django-redis